Work on 1.4A
Technically ready for manufacture, but I want to play around with it a bit more first. Electrolytic capacitor swapped to SMD, new capacitor saves on height and decreases profile of board. This should make it easier to mount and not worry about collisions etc., board is less likely to be damaged.